# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Private PMC membership and contribution platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /member/{memberId}:
    get:
      operationId: getPublicMemberProfile
      tags: [public]
      summary: Get the safe public profile linked from a member QR code
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leader/department:
    get:
      operationId: getLeaderDepartment
      tags: [leader]
      summary: Get the leader's department and its member contribution summary
      parameters:
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Department data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentSummary"
        "403":
          description: Leader permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leader/credits:
    post:
      operationId: awardCredit
      tags: [leader]
      summary: Award a predefined credit to an eligible member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditAwardInput"
      responses:
        "201":
          description: Created credit transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditTransaction"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leader/achievements:
    post:
      operationId: createAchievement
      tags: [leader]
      summary: Create an achievement for an eligible member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementInput"
      responses:
        "201":
          description: Created achievement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        errorAr:
          type: ["string", "null"]
...
    CurrentUser:
      type: object
      required: [id, memberId, fullName, email, role, membershipStatus, credits, level]
      properties:
        id: { type: string }
        memberId: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        college: { type: ["string", "null"] }
        major: { type: ["string", "null"] }
        role:
          type: string
          enum: [MEMBER, DEPARTMENT_LEADER, PRESIDENT]
        membershipStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED, INACTIVE]
        department:
          type: ["string", "null"]
        credits: { type: integer }
        level:
          $ref: "#/components/schemas/LevelSummary"
        joinedAt: { type: string, format: date-time }
        emailVerified: { type: boolean }
    LevelSummary:
      type: object
      required: [key, nameEn, nameAr, symbol, current, nextThreshold, progressPercent]
      properties:
        key: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        symbol: { type: string }
        current: { type: integer }
        nextThreshold: { type: ["integer", "null"] }
        progressPercent: { type: number }
...
    CreditType:
      type: object
      required: [id, nameAr, nameEn, creditValue, active, requiresPresidentApproval]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: string }
        creditValue: { type: integer }
        active: { type: boolean }
        requiresPresidentApproval: { type: boolean }
        description: { type: ["string", "null"] }
    CreditTransaction:
      type: object
      required: [id, memberId, memberName, creditType, creditValue, activityName, activityType, awardedBy, createdAt]
      properties:
        id: { type: string }
        memberId: { type: string }
        memberName: { type: string }
        creditType:
          $ref: "#/components/schemas/CreditType"
        creditValue: { type: integer }
        activityName: { type: string }
        activityType: { type: string }
        awardedBy: { type: string }
        approvedBy: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CreditAwardInput:
      type: object
      required: [memberId, creditTypeId, activityName, activityDate]
      properties:
        memberId: { type: string }
        creditTypeId: { type: integer }
        activityName: { type: string, minLength: 2 }
        activityDate: { type: string, format: date }
        note: { type: ["string", "null"] }
    Achievement:
      type: object
      required: [id, memberId, type, title, activity, date, approvedBy]
      properties:
        id: { type: string }
        memberId: { type: string }
        type:
          type: string
          enum: [PROJECT, CHALLENGE, LEADERSHIP, EVENT, IDEA, CONTRIBUTOR, SPECIAL]
        title: { type: string }
        activity: { type: string }
        date: { type: string, format: date }
        approvedBy: { type: string }
        description: { type: ["string", "null"] }
    AchievementInput:
      type: object
      required: [memberId, type, title, activity, date]
      properties:
        memberId: { type: string }
        type:
          type: string
          enum: [PROJECT, CHALLENGE, LEADERSHIP, EVENT, IDEA, CONTRIBUTOR, SPECIAL]
        title: { type: string, minLength: 2 }
        activity: { type: string, minLength: 2 }
        date: { type: string, format: date }
        description: { type: ["string", "null"] }
...
    PublicProfile:
      type: object
      required: [memberId, name, level, credits, achievements, memberSince]
      properties:
        memberId: { type: string }
        name: { type: string }
        level: { $ref: "#/components/schemas/LevelSummary" }
        credits: { type: integer }
        achievements:
          type: array
          items: { $ref: "#/components/schemas/Achievement" }
        memberSince: { type: string, format: date-time }
        department: { type: ["string", "null"] }
    Department:
      type: object
      required: [id, nameAr, nameEn, active, memberCount]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: string }
        active: { type: boolean }
        memberCount: { type: integer }
        leader:
          type: ["object", "null"]
          properties:
            memberId: { type: string }
            name: { type: string }
...
    MemberRecord:
      allOf:
        - $ref: "#/components/schemas/CurrentUser"
        - type: object
          required: [accountActivated, active]
          properties:
            accountActivated: { type: boolean }
            active: { type: boolean }
...
    DepartmentSummary:
      type: object
      required: [department, members]
      properties:
        department: { $ref: "#/components/schemas/Department" }
        members:
          type: array
          items: { $ref: "#/components/schemas/MemberRecord" }
...